Day4/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
PyJWT==2.9.0
itsdangerous==2.2.0